Compiles and runs now. Need to adjust MPI parallelization in "ParallelSimParser.cpp", consider removing "ThreadPool.h"

+++ Potentials.h
@@ -100,7 +100,7 @@ namespace Envelopes {
 	private:
 		double tau;
 		double tmax;
-		double a_t = 2.0 * std::acos(std::powf(0.5, 0.25)); // cos^4(a_t*(tau/2)/tau)=1/2 s.t. tau = FWHM-power
+		double a_t = 2.0 * std::acos(std::pow(0.5, 0.25)); // cos^4(a_t*(tau/2)/tau)=1/2 s.t. tau = FWHM-power
 	public:
 		CosSquaredEnvelope(double tau, double tmax);
 		double getValue(double t);
